demo(android): unify animation + physics into animation-physics (#2239 Batch 3)
Consolidates the retired `animation` and `physics` demos into a single
`animation-physics` entry with a 2-way segmented-button toggle (mirrors
#2239 Batch 1/2):

- **Animation** (default) — skeletal/keyframe playback: model carousel,
  cinematic camera shots (Hero/Reveal/Vertigo/Tracking/Free), and
  play/pause/speed/loop controls, plus the qaMode/stopAnimation freeze.
- **Physics** — PhysicsNode rigid-body simulation: drop streamed
  crash-test bodies or bundled spheres that fall under gravity and bounce.

Each sub-mode owns its own rememberEngine + loaders + SceneView so
switching tabs tears down the inactive section completely (prevents
resource leaks). All QA-day / #issue rationale comments preserved from
the source demos.

Retired deep-link ids (`animation`, `physics`) keep working via
DeepLinkRouter.DEMO_ID_ALIASES (and the RootScreen Explore-tab curated
list now points at `animation-physics`).

Tests: DemoInteractionTest.animationPhysics_allTabs taps both tabs;
DemoRenderingScreenshotTest.animationPhysicsDemo_default_state (golden
animationphysics_default). Stale animation_default.png golden removed.
llms.txt demos block + GeneratedDemos.kt regenerated via collate-demos.sh.
