+++ changelog.d/2239-batch-3-animation-physics.md
@@ -1,0 +1,2 @@
+<!-- category: Changed -->
+- **Samples catalog: unified Animation & Physics demo ([#2239](https://github.com/sceneview/sceneview/issues/2239)).** The retired `animation` and `physics` demos consolidated into a single `animation-physics` entry with a segmented-button toggle between Animation (skeletal/keyframe playback with a model carousel, cinematic camera shots, and play/pause/speed/loop controls) and Physics (`PhysicsNode` rigid-body simulation dropping streamed crash-test bodies or bundled spheres). Existing `sceneview://demo/animation|physics` deep links keep working via `DEMO_ID_ALIASES`.
