+++ samples/android-demo/src/main/java/io/github/sceneview/demo/demos/AnimationPhysicsDemo.kt
@@ -17,33 +17,40 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Pause
 import androidx.compose.material.icons.filled.PlayArrow
+import androidx.compose.material3.Button
 import androidx.compose.material3.FilterChip
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SegmentedButton
+import androidx.compose.material3.SegmentedButtonDefaults
+import androidx.compose.material3.SingleChoiceSegmentedButtonRow
 import androidx.compose.material3.Slider
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.key
 import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.produceState
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.compose.runtime.produceState
-import androidx.compose.ui.platform.LocalContext
+import com.google.android.filament.LightManager
 import io.github.sceneview.ExperimentalSceneViewApi
 import io.github.sceneview.SceneView
 import io.github.sceneview.demo.DemoScaffold
-import io.github.sceneview.demo.R
 import io.github.sceneview.demo.DemoSettings
 import io.github.sceneview.demo.LoadingScrim
+import io.github.sceneview.demo.R
+import io.github.sceneview.demo.SceneViewColors
 import io.github.sceneview.demo.demos.internal.DemoMath
 import io.github.sceneview.demo.rememberFirstFrameState
 import io.github.sceneview.demo.sketchfab.SampleAssets
@@ -51,51 +58,111 @@ import io.github.sceneview.demo.sketchfab.SketchfabAssetResolver
 import io.github.sceneview.demo.sketchfab.SketchfabSlug
 import io.github.sceneview.environment.rememberHDREnvironment
 import io.github.sceneview.gesture.CameraGestureDetector
+import io.github.sceneview.loaders.ModelLoader
 import io.github.sceneview.math.Position
+import io.github.sceneview.math.Size
 import io.github.sceneview.math.Transform
+import io.github.sceneview.model.Model
 import io.github.sceneview.node.ModelNode as ModelNodeImpl
+import io.github.sceneview.node.SphereNode as SphereNodeImpl
+import io.github.sceneview.rememberCameraManipulator
 import io.github.sceneview.rememberCameraNode
 import io.github.sceneview.rememberEngine
 import io.github.sceneview.rememberEnvironment
 import io.github.sceneview.rememberEnvironmentLoader
+import io.github.sceneview.rememberMaterialLoader
 import io.github.sceneview.rememberModelInstance
 import io.github.sceneview.rememberModelLoader
 import io.github.sceneview.sample.LifecyclePausingLaunchedEffect
+import io.github.sceneview.sample.rememberMaterialInstance
 import java.io.File
 import kotlin.math.cos
 import kotlin.math.sin
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
 /**
- * Demonstrates model animation playback controls: play/pause, speed, and loop mode.
+ * Unified "Animation & Physics" demo — consolidates the retired `animation` and
+ * `physics` demos behind a single segmented-button toggle (#2239 Batch 3).
  *
- * The `autoAnimate` parameter on `ModelNode` is only read once at node creation, and
- * the composable's reactive `animationName` path doesn't re-key on speed/loop
- * changes — so we drive the animation state imperatively through a `LaunchedEffect`
- * that watches (isPlaying, speed, loop) and calls `playAnimation` / `stopAnimation`
- * on a captured node reference. That gives the three chips a real effect.
+ * - **Animation** (default) — skeletal / keyframe animation playback with a model
+ *   carousel, cinematic camera shots, and play / pause / speed / loop controls.
+ *   (Formerly `animation`.)
+ * - **Physics** — rigid-body simulation: drop streamed crash-test bodies (or
+ *   bundled spheres) that fall under gravity and bounce off the floor. (Formerly
+ *   `physics`.)
  *
- * Cinematic camera (Phase 3 — real cinematic shots)
- * --------------------------------------------------
- * The four scripted modes are not generic primitives ("orbit", "dolly", "crane") any
- * more — each one is a real cinematic shot type with its own keyframed choreography:
- *
- *   - HERO    — slow heroic low-angle orbit with a 2 s pause at the front-3/4 hold
- *   - REVEAL  — close-up chest framing pulls back to a slight high-angle wide shot
- *   - VERTIGO — Hitchcock dolly-zoom (radius and FOV move in opposite directions)
- *   - TRACKING — straight-line lateral pass with a per-frame lookAt back at the subject
- *   - FREE    — user gesture only (DefaultCameraManipulator), no scripted motion
- *
- * The single [ScriptedCameraManipulator] is parameterized on lambdas (yaw, radius,
- * yHeight, *and* a full-eye override) so TRACKING can leave the orbit circle and
- * move along a straight line while still re-aiming at the subject every frame.
- *
- * FOV control (for the dolly-zoom shot) goes through the SDK's [CameraNode] —
- * we own the camera node via [rememberCameraNode] and call `setProjection(fov, ...)`
- * each composition pass, driven by an [Animatable]. Since `setProjection` clamps
- * silently (0 < fov < 180) and only no-ops while `view == null`, this is safe to call
- * before the first frame.
+ * Each sub-mode owns its own `SceneView` + its own [rememberEngine] / loaders,
+ * so switching tabs tears down the inactive section completely — no engine is
+ * hoisted above the `when`, which is what prevents resource leaks across tab
+ * switches (Batch 1 review confirmed this pattern). Old deep links route
+ * through [io.github.sceneview.demo.DeepLinkRouter.DEMO_ID_ALIASES].
  */
+@Composable
+fun AnimationPhysicsDemo(onBack: () -> Unit) {
+    var mode by remember { mutableStateOf(AnimationPhysicsMode.Animation) }
+    when (mode) {
+        AnimationPhysicsMode.Animation -> AnimationSection(onBack, mode) { mode = it }
+        AnimationPhysicsMode.Physics -> PhysicsSection(onBack, mode) { mode = it }
+    }
+}
+
+private enum class AnimationPhysicsMode(val label: String) {
+    Animation("Animation"),
+    Physics("Physics"),
+}
+
+@Composable
+private fun ModeSelector(
+    current: AnimationPhysicsMode,
+    onModeChange: (AnimationPhysicsMode) -> Unit,
+) {
+    val modes = AnimationPhysicsMode.entries
+    SingleChoiceSegmentedButtonRow(modifier = Modifier.fillMaxWidth()) {
+        modes.forEachIndexed { index, m ->
+            SegmentedButton(
+                selected = m == current,
+                onClick = { onModeChange(m) },
+                shape = SegmentedButtonDefaults.itemShape(index = index, count = modes.size),
+                label = { Text(m.label) },
+            )
+        }
+    }
+    Spacer(modifier = Modifier.height(12.dp))
+}
+
+// ─── Animation section ──────────────────────────────────────────────────────
+// Formerly AnimationDemo.
+//
+// Demonstrates model animation playback controls: play/pause, speed, and loop mode.
+//
+// The `autoAnimate` parameter on `ModelNode` is only read once at node creation, and
+// the composable's reactive `animationName` path doesn't re-key on speed/loop
+// changes — so we drive the animation state imperatively through a `LaunchedEffect`
+// that watches (isPlaying, speed, loop) and calls `playAnimation` / `stopAnimation`
+// on a captured node reference. That gives the three chips a real effect.
+//
+// Cinematic camera (Phase 3 — real cinematic shots)
+// --------------------------------------------------
+// The four scripted modes are not generic primitives ("orbit", "dolly", "crane") any
+// more — each one is a real cinematic shot type with its own keyframed choreography:
+//
+//   - HERO    — slow heroic low-angle orbit with a 2 s pause at the front-3/4 hold
+//   - REVEAL  — close-up chest framing pulls back to a slight high-angle wide shot
+//   - VERTIGO — Hitchcock dolly-zoom (radius and FOV move in opposite directions)
+//   - TRACKING — straight-line lateral pass with a per-frame lookAt back at the subject
+//   - FREE    — user gesture only (DefaultCameraManipulator), no scripted motion
+//
+// The single [ScriptedCameraManipulator] is parameterized on lambdas (yaw, radius,
+// yHeight, *and* a full-eye override) so TRACKING can leave the orbit circle and
+// move along a straight line while still re-aiming at the subject every frame.
+//
+// FOV control (for the dolly-zoom shot) goes through the SDK's [CameraNode] —
+// we own the camera node via [rememberCameraNode] and call `setProjection(fov, ...)`
+// each composition pass, driven by an [Animatable]. Since `setProjection` clamps
+// silently (0 < fov < 180) and only no-ops while `view == null`, this is safe to call
+// before the first frame.
 private enum class CameraMode { HERO, REVEAL, VERTIGO, TRACKING, FREE }
 
 /**
@@ -132,7 +199,7 @@ private data class AnimationModel(
 }
 
 /**
- * Carousel of 5 animated models for [AnimationDemo].
+ * Carousel of 5 animated models for [AnimationSection].
  *
  * Slot 0 is the historical `threejs_soldier.glb` (bundled, 4 animations:
  * 0=Idle, 1=Run, 2=TPose, 3=Walk). The next 4 slots stream the 4 entries from
@@ -185,7 +252,11 @@ private val ANIMATION_MODELS: List<AnimationModel> = run {
 
 @OptIn(ExperimentalSceneViewApi::class)
 @Composable
-fun AnimationDemo(onBack: () -> Unit) {
+private fun AnimationSection(
+    onBack: () -> Unit,
+    mode: AnimationPhysicsMode,
+    onModeChange: (AnimationPhysicsMode) -> Unit,
+) {
     // Index into [ANIMATION_MODELS] — defaults to the historical soldier so the
     // first frame looks identical to v4.3.1 when the carousel is unused.
     var selectedModelIndex by remember { mutableIntStateOf(0) }
@@ -604,10 +675,11 @@ fun AnimationDemo(onBack: () -> Unit) {
     val firstFrame = rememberFirstFrameState()
 
     DemoScaffold(
-        title = stringResource(R.string.demo_animation_title),
+        title = stringResource(R.string.demo_animation_physics_title),
         onBack = onBack,
         firstFrameRendered = firstFrame.rendered,
         controls = {
+            ModeSelector(mode, onModeChange)
             // Model carousel — switches the active animated subject. Slot 0 is
             // the bundled threejs soldier; slots 1–4 stream from the `animation`
             // category of SampleAssets. Streamed slots fall back to the bundled
@@ -643,13 +715,13 @@ fun AnimationDemo(onBack: () -> Unit) {
                 horizontalArrangement = Arrangement.spacedBy(8.dp),
                 verticalAlignment = Alignment.CenterVertically
             ) {
-                CameraMode.entries.forEach { mode ->
+                CameraMode.entries.forEach { camMode ->
                     FilterChip(
-                        selected = cameraMode == mode,
-                        onClick = { cameraMode = mode },
+                        selected = cameraMode == camMode,
+                        onClick = { cameraMode = camMode },
                         label = {
                             Text(
-                                when (mode) {
+                                when (camMode) {
                                     CameraMode.HERO -> "Hero"
                                     CameraMode.REVEAL -> "Reveal"
                                     CameraMode.VERTIGO -> "Vertigo"
@@ -801,7 +873,7 @@ fun AnimationDemo(onBack: () -> Unit) {
 /**
  * Records an in-flight gesture's begin event so the host can replay it on the
  * fresh `DefaultCameraManipulator` after the FREE-mode swap. See [freeManipulator]
- * in [AnimationDemo] for the rationale (TL;DR — without replay the new manipulator
+ * in [AnimationSection] for the rationale (TL;DR — without replay the new manipulator
  * has no origin point and every grabUpdate produces zero deltas).
  */
 private sealed class PendingGestureBegin {
@@ -889,3 +961,321 @@ private class ScriptedCameraManipulator(
     @Suppress("EmptyFunctionBlock")
     override fun update(deltaTime: Float) {}
 }
+
+// ─── Physics section ────────────────────────────────────────────────────────
+// Formerly PhysicsDemo.
+//
+// Demonstrates [PhysicsNode] — drop streamed crash-test bodies (chairs, vases,
+// barrels, amphorae) that fall under gravity and bounce off the floor.
+//
+// Stage 2 migration ([#1152](https://github.com/sceneview/sceneview/issues/1152)).
+//  - Previous version: 5 coloured spheres bouncing on a plane — useful to verify the
+//    rigid-body integrator but not a real visual showcase of "physics on a real GLB".
+//  - New version: the same simulation, but with the four streamed entries from
+//    [SampleAssets.byCategory]`["physics"]` (Ceramic Vase, Wooden Stool,
+//    Wooden Barrel, Clay Amphora — all CC-BY from Sketchfab) cycling through each
+//    drop. The "Bundled spheres" chip preserves the v4.3.1 spheres-only mode for
+//    QA / offline / store-listing screenshot determinism.
+//
+// Physics math is unchanged — every dropped object is treated as a sphere of
+// `collisionRadius = 0.1 m` so the bounce reads naturally regardless of the actual
+// mesh shape (we don't ship a convex-hull collider — `feedback_demo_quality` calls
+// out that the demo's goal is to showcase the SDK's wiring, not to ship a physics
+// engine). The visual mesh is a `ModelNode` parented to the simulated [SphereNodeImpl];
+// the parent sphere is rendered invisibly via a transparent material so only the
+// mesh is visible to the viewer.
+//
+// Each "Drop" press adds a new body. "Reset" clears every body by incrementing a
+// generation key that forces full recomposition. Streamed slugs use the resolver's
+// fallback path when no Sketchfab key is configured, so the carousel always drops
+// something visible even offline.
+@Composable
+private fun PhysicsSection(
+    onBack: () -> Unit,
+    mode: AnimationPhysicsMode,
+    onModeChange: (AnimationPhysicsMode) -> Unit,
+) {
+    // Start with 5 bodies so the first frame already shows the demo's hook
+    // (a colourful rain on the floor) instead of a near-empty scene.
+    var bodyCount by remember { mutableIntStateOf(5) }
+    var generation by remember { mutableIntStateOf(0) }
+
+    // Streamed `physics` slugs from SampleAssets. selectedSlug == null means
+    // "Bundled spheres" — the v4.3.1 visual default. Selecting a slug arms
+    // it as the carousel of streamed crash-test bodies (chairs / vases /
+    // barrels / amphorae) cycling through each drop.
+    val physicsSlugs = remember { SampleAssets.byCategory["physics"].orEmpty() }
+    var selectedSlug by remember { mutableStateOf<SketchfabSlug?>(physicsSlugs.firstOrNull()) }
+
+    val context = LocalContext.current
+
+    // Warm the `physics` cache so the very first drop renders without a pop-in.
+    // The resolver dedupes concurrent calls, so the per-body resolve below picks
+    // up the cached file as soon as the prefetch lands.
+    LaunchedEffect(Unit) {
+        runCatching {
+            SketchfabAssetResolver.getInstance(context).prefetchAll("physics")
+        }
+    }
+
+    // Resolve the currently-selected slug to a local file (null while
+    // downloading / staging the bundled fallback). When null, drops fall
+    // back to the original spheres-only mode so the user sees something
+    // moving while the streamed mesh lands.
+    val selectedFile: File? = selectedSlug?.let { slug ->
+        produceState<File?>(initialValue = null, key1 = slug.uid) {
+            value = runCatching {
+                SketchfabAssetResolver.getInstance(context).resolve(slug)
+            }.getOrNull()
+        }.value
+    }
+
+    val engine = rememberEngine()
+    val modelLoader = rememberModelLoader(engine)
+    val materialLoader = rememberMaterialLoader(engine)
+    val environmentLoader = rememberEnvironmentLoader(engine)
+    // Camera above and back from the scene, angled down. The look-at target sits between
+    // the floor (y = -0.5) and the spheres' rest height (y ≈ -0.42) rather than the scene
+    // origin, so the 1.6 m ground plane is vertically centred in the viewport instead of
+    // being shoved into the bottom third with its near edge clipped (#1463). Pulled back to
+    // z = 4 so the full plane depth fits with headroom for the drop column above it.
+    val cameraNode = rememberCameraNode(engine) {
+        position = Position(0f, 2f, 4f)
+        lookAt(Position(0f, -0.35f, 0f))
+    }
+
+    val firstFrame = rememberFirstFrameState()
+
+    DemoScaffold(
+        title = stringResource(R.string.demo_animation_physics_title),
+        onBack = onBack,
+        firstFrameRendered = firstFrame.rendered,
+        controls = {
+            ModeSelector(mode, onModeChange)
+            Text("Bodies: $bodyCount", style = MaterialTheme.typography.labelLarge)
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.spacedBy(8.dp)
+            ) {
+                Button(onClick = { bodyCount++ }) {
+                    Text("Drop")
+                }
+                Button(onClick = { bodyCount += 10 }) {
+                    Text("Drop 10")
+                }
+                Button(onClick = {
+                    bodyCount = 1
+                    generation++
+                }) {
+                    Text("Reset")
+                }
+            }
+
+            Spacer(modifier = Modifier.height(12.dp))
+            Text(
+                text = stringResource(R.string.demo_physics_picker_label),
+                style = MaterialTheme.typography.labelLarge,
+            )
+            Spacer(modifier = Modifier.height(4.dp))
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .horizontalScroll(rememberScrollState()),
+                horizontalArrangement = Arrangement.spacedBy(8.dp),
+            ) {
+                // "Bundled spheres" chip preserves the v4.3.1 visual default
+                // — useful for QA / offline / store-listing screenshots.
+                FilterChip(
+                    selected = selectedSlug == null,
+                    onClick = {
+                        selectedSlug = null
+                        // Reset so the chip swap is unambiguous — spheres
+                        // first, then more spheres on tap.
+                        bodyCount = 5
+                        generation++
+                    },
+                    label = {
+                        Text(stringResource(R.string.demo_physics_picker_spheres))
+                    },
+                )
+                physicsSlugs.forEach { slug ->
+                    FilterChip(
+                        selected = selectedSlug?.uid == slug.uid,
+                        onClick = {
+                            selectedSlug = slug
+                            bodyCount = 5
+                            generation++
+                        },
+                        label = { Text(slug.displayName) },
+                    )
+                }
+            }
+            Spacer(modifier = Modifier.height(4.dp))
+            Text(
+                text = stringResource(R.string.demo_physics_picker_subtitle),
+                style = MaterialTheme.typography.labelSmall,
+            )
+        }
+    ) {
+        // key(generation) forces full recomposition on reset
+        key(generation) {
+            SceneView(
+                modifier = Modifier.fillMaxSize(),
+                engine = engine,
+                modelLoader = modelLoader,
+                materialLoader = materialLoader,
+                environmentLoader = environmentLoader,
+                cameraNode = cameraNode,
+                onFrame = firstFrame.onFrame,
+                cameraManipulator = rememberCameraManipulator(
+                    orbitHomePosition = cameraNode.worldPosition
+                )
+            ) {
+                // Left-side counter-fill — same as v4.3.1, kept verbatim.
+                LightNode(
+                    type = LightManager.Type.DIRECTIONAL,
+                    direction = io.github.sceneview.math.Direction(-0.3f, -1f, -0.5f),
+                    apply = {
+                        intensity(5_000f)
+                    }
+                )
+                val groundMaterial = rememberMaterialInstance(
+                    materialLoader, SceneViewColors.SurfaceDim
+                )
+                // Ramp4 is a fixed 4-colour list — call the helper once per slot so
+                // each MaterialInstance gets the same disposal hygiene as the rest.
+                val sphereMaterials = listOf(
+                    rememberMaterialInstance(materialLoader, SceneViewColors.Ramp4[0]),
+                    rememberMaterialInstance(materialLoader, SceneViewColors.Ramp4[1]),
+                    rememberMaterialInstance(materialLoader, SceneViewColors.Ramp4[2]),
+                    rememberMaterialInstance(materialLoader, SceneViewColors.Ramp4[3]),
+                )
+
+                // Ground plane — must use Size(x, y=0, z) for a HORIZONTAL floor
+                PlaneNode(
+                    materialInstance = groundMaterial,
+                    size = Size(x = 1.6f, y = 0f, z = 1.6f),
+                    position = Position(y = -0.5f),
+                )
+
+                // Streamed mesh path. The model file is `null` until the
+                // resolver returns (or the user picked "Bundled spheres").
+                // The GLB is parsed once into a single `Model` (geometry +
+                // materials live here, shared by every instance); each falling
+                // body then gets its OWN `ModelInstance` spawned from it below.
+                // A `ModelInstance` wraps exactly one Filament entity / one
+                // TransformManager slot, so it can only ride one body at a
+                // time — sharing it across bodies meant every ModelNode wrote
+                // the same entity transform (last-write-wins) and only one
+                // mesh was ever visible (#1706).
+                val streamedModel: Model? = selectedFile?.let { file ->
+                    rememberStreamedModel(modelLoader, file)
+                }
+
+                // collisionRadius is the bouncing-sphere radius PhysicsNode uses
+                // to offset the contact point off the floor. Keep it consistent
+                // regardless of whether we render a sphere or a streamed mesh —
+                // the demo's value is the simulation hook-up, not a per-mesh
+                // collider, and `feedback_demo_quality` says the SDK is the
+                // showcase, not bespoke physics.
+                val collisionRadius = 0.08f
+
+                for (i in 0 until bodyCount) {
+                    val xOffset = (i % 5 - 2) * 0.18f + (if (i / 5 % 2 == 0) 0f else 0.09f)
+                    val zOffset = ((i / 5) % 3 - 1) * 0.18f
+                    val startY = 0.6f + (i / 5) * 0.18f
+
+                    var nodeRef by remember(i) { mutableStateOf<SphereNodeImpl?>(null) }
+
+                    // The simulated SphereNode is rendered "invisibly" (it
+                    // carries the colour ramp material when the user is in
+                    // bundled-sphere mode; in streamed mode we ALSO render
+                    // it — same coloured silhouette — so the dropped streamed
+                    // mesh sits visually on top of a soft colour pad which
+                    // hides the bounding-sphere abstraction).
+                    SphereNode(
+                        radius = collisionRadius,
+                        materialInstance = sphereMaterials[i % 4],
+                        position = Position(x = xOffset, y = startY, z = zOffset),
+                        apply = { nodeRef = this }
+                    ) {
+                        // Streamed mesh child — only rendered when a streamed
+                        // slug is selected AND its download has landed. The
+                        // child inherits the sphere's transform so it rides
+                        // the simulation. Each body spawns its OWN
+                        // `ModelInstance` from the shared `Model` so it has an
+                        // independent Filament entity — `createInstance` only
+                        // duplicates the lightweight entity tree, geometry and
+                        // materials stay shared (#1706). Keyed on `i` so the
+                        // instance is created once per body, on the main
+                        // composition thread (Filament JNI is @MainThread).
+                        val model = streamedModel
+                        val slug = selectedSlug
+                        if (model != null && slug != null) {
+                            val instance = remember(i, model) {
+                                modelLoader.createInstance(model)
+                            }
+                            if (instance != null) {
+                                ModelNode(
+                                    modelInstance = instance,
+                                    scaleToUnits = slug.scaleToUnits,
+                                    centerOrigin = Position(0f, 0f, 0f),
+                                )
+                            }
+                        }
+                    }
+
+                    // PhysicsNode attaches an onFrame callback that applies
+                    // gravity + bounce. The radius is the bounding-sphere
+                    // collision radius — the streamed mesh child rides
+                    // visually on top.
+                    nodeRef?.let { node ->
+                        PhysicsNode(
+                            node = node,
+                            restitution = 0.7f,
+                            floorY = -0.5f,
+                            radius = collisionRadius,
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
+
+/**
+ * Parses the streamed GLB at [file] once into a single [Model] that every
+ * dropped body then spawns its own [ModelInstance] from.
+ *
+ * `releaseSourceData = false` is mandatory here: [ModelLoader.createInstance]
+ * cannot run after the source glTF data has been released, so we keep it
+ * resident for as long as new bodies may still be dropped.
+ *
+ * Threading mirrors `rememberModelInstance`: the file bytes are read on
+ * [Dispatchers.IO], then `createModel` (a `@MainThread` Filament JNI call)
+ * runs back on the composition's main dispatcher inside [produceState].
+ * Returns `null` while loading.
+ */
+@Composable
+private fun rememberStreamedModel(
+    modelLoader: ModelLoader,
+    file: File,
+): Model? = produceState<Model?>(initialValue = null, key1 = modelLoader, key2 = file.absolutePath) {
+    // Read the GLB bytes (and any external glTF resources) off the main
+    // thread, then call `createModel` — a @MainThread Filament JNI call —
+    // back on the composition's main dispatcher (produceState's context).
+    val buffer = withContext(Dispatchers.IO) {
+        runCatching { java.nio.ByteBuffer.wrap(file.readBytes()) }.getOrNull()
+    } ?: return@produceState
+    value = runCatching {
+        modelLoader.createModel(
+            buffer = buffer,
+            releaseSourceData = false,
+            resourceResolver = { resourceFile ->
+                runCatching {
+                    java.nio.ByteBuffer.wrap(File(file.parent, resourceFile).readBytes())
+                }.getOrNull()
+            },
+        )
+    }.getOrNull()
+}.value
