--- samples/android-demo/src/main/res/values/strings_demo_animation.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `animation` demo (registered by
-     `AnimationFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_animation_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_animation_title">Animation</string>
-    <string name="demo_animation_subtitle">Play, pause, and control animations</string>
-    <string name="demo_animation_control">Animation Control</string>
-    <string name="demo_animation_control_desc">Switch animations by name, control speed and looping</string>
-</resources>
