--- samples/android-demo/src/main/res/values/strings_demo_physics.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `physics` demo (registered by
-     `PhysicsFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_physics_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_physics_title">Physics</string>
-    <string name="demo_physics_subtitle">Gravity, collisions, and rigid bodies</string>
-    <string name="demo_physics">Physics</string>
-    <string name="demo_physics_desc">Rigid body simulation with bouncing balls</string>
-    <string name="demo_physics_picker_label">What to drop</string>
-    <string name="demo_physics_picker_spheres">Bundled spheres</string>
-    <string name="demo_physics_picker_subtitle">Crash-test bodies are CC-BY models from Sketchfab. Switching the picker resets the scene so the new shape is what falls.</string>
-</resources>
